+++ Universal/ModernDisplayEngineDxe/FormDisplay.c
@@ -2423,8 +2423,7 @@ DisplayOneMenu (
   CHAR16                         AdjustValue;
   UINTN                          MaxRow;
   UINTN                          RowWidth;
-  BOOLEAN                        IsActionRow;
-  BOOLEAN                        IsSubtitleRow;
+
 
   Statement            = MenuOption->ThisTag;
   Temp                 = SkipLine;
@@ -2435,26 +2434,19 @@ DisplayOneMenu (
   OptionLineNum        = 0;
   MaxRow               = 0;
   IsProcessingFirstRow = TRUE;
-  IsActionRow          = (BOOLEAN)(
-                                  (Statement->OpCode->OpCode == EFI_IFR_ACTION_OP) ||
-                                  (Statement->OpCode->OpCode == EFI_IFR_REF_OP) ||
-                                  (Statement->OpCode->OpCode == EFI_IFR_RESET_BUTTON_OP)
-                                  );
-  IsSubtitleRow        = (BOOLEAN)(Statement->OpCode->OpCode == EFI_IFR_SUBTITLE_OP);
-
   //
   // Set default color.
   //
   SetDisplayAttribute (MenuOption, FALSE);
   RowWidth = (gPromptBlockWidth + gOptionBlockWidth + LEFT_SKIPPED_COLUMNS);
   ModernDisplayDrawStatementRow (
+    gFormData,
+    Statement,
     BeginCol,
     MenuOption->Row,
     RowWidth,
     Highlight,
-    MenuOption->GrayOut,
-    IsActionRow,
-    IsSubtitleRow
+    (BOOLEAN)(gUserInput->SelectedStatement == Statement)
     );
 
   //